rename series repository package name
