+++ data/remote_data_source/series/src/test/java/com/sanaa/series/RemoteTvSeriesDataSourceImplTest.kt
@@ -1,7 +1,6 @@
 package com.sanaa.series
 
 import com.sanaa.preferences.service.LanguageProvider
-import com.sanaa.series.data_source.remote.RemoteTvSeriesDataSource
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.mock.MockEngine
 import io.ktor.client.engine.mock.MockRequestHandleScope
